+++ plugins/ci/pkg/ci/trivy.go
@@ -12,12 +12,12 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/fairwindsops/insights-plugins/trivy/pkg/image"
-	trivymodels "github.com/fairwindsops/insights-plugins/trivy/pkg/models"
+	"github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/image"
+	trivymodels "github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/models"
 	"github.com/sirupsen/logrus"
 
-	"github.com/fairwindsops/insights-plugins/ci/pkg/commands"
-	"github.com/fairwindsops/insights-plugins/ci/pkg/models"
+	"github.com/fairwindsops/insights-plugins/plugins/ci/pkg/commands"
+	"github.com/fairwindsops/insights-plugins/plugins/ci/pkg/models"
 )
 
 func (ci *CIScan) GetTrivyReport(manifestImages []trivymodels.Image) (models.ReportInfo, error) {