+++ plugins/right-sizer/src/report/report.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/fairwindsops/insights-plugins/right-sizer/src/util"
+	"github.com/fairwindsops/insights-plugins/plugins/right-sizer/src/util"
 	"github.com/golang/glog"
 	core "k8s.io/api/core/v1"
 	kube_errors "k8s.io/apimachinery/pkg/api/errors"